third_party/WebKit/LayoutTests/http/tests/loading/sxg/fallback-to-another-sxg.html

<!DOCTYPE html>
<title>Fallback redirect to another SignedHTTPExchange</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  await waitUntilDidFinishLoadForFrame;

  // The timestamp of the test SXG file is "Apr 1 2018 00:00 UTC" and valid
  // until "Apr 8 2018 00:00 UTC".
  await setSignedExchangeVerificationTime(new Date("Apr 1 2018 00:01 UTC"));

  const message = await openSXGInIframeAndWaitForMessage(t, 'resources/fallback-to-another-sxg.sxg');
  assert_equals(message.location, 'https://127.0.0.1:8443/loading/sxg/resources/inner-url.html');
  assert_false(message.is_fallback);
}, 'Fallback redirect to another SignedHTTPExchange');

</script>
</body>
